% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_expand_colnames}
\alias{academic_libraries_expand_colnames}
\title{Converts column short names in an AL dataset into long names.}
\usage{
academic_libraries_expand_colnames(data)
}
\arguments{
\item{data}{The academic libraries file loaded from `academic_libraries_load_file`.}
}
\value{
The Academic Libraries file with expanded variable names.
}
\description{
Changes the variable names in an Academic Libraries file to the more descriptive
             versions listed in the latest AL dictionary file.
}
\note{
This function should be called after joining multiple AL files.
}
\examples{
library(tidyverse)

# Assuming you have downloaded the AL files
# (e.g., "al2018_rv.csv", "al2018.xlsx") from
#  2014-2019 into your documents folder.
set_ipeds_data_folder(path.expand('~'))
  
ALData <-
  academic_libraries_load_file(year=2014,
                               institutions=oh_community_colleges) \%>\%
    union_all(academic_libraries_load_file(year=2015)) \%>\%
    union_all(academic_libraries_load_file(year=2016)) \%>\%
    union_all(academic_libraries_load_file(year=2017)) \%>\%
    union_all(academic_libraries_load_file(year=2018)) \%>\%
    union_all(academic_libraries_load_file(year=2019)) \%>\%
    academic_libraries_expand_colnames()
}
\author{
Blake Madden
}
