% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_load_file}
\alias{academic_libraries_load_file}
\title{Loads and filters an Academic Libraries file.}
\usage{
academic_libraries_load_file(year, ..., institutions = NULL)
}
\arguments{
\item{year}{The year in the file name to load.}

\item{...}{Institutions to collapse. This is an argument to `fct_collapse`, which is an
institution and its character vector of satellite institutions to collapse into it.
An example could be `"206604" = c("206613")`, and you can use a variadic list of assignments such as these.
Refer to the documentation for `fct_collapse` for further details.}

\item{institutions}{A vector of college IDs (i.e., the UNITID) to filter on.}
}
\value{
The transformed Academic Libraries file.
}
\description{
Loads an Academic Libraries file, removes the imputation columns,
             and optionally filters and combines institutions.
}
\note{
Additional functions to pivot this data are available.
}
\examples{
library(tidyverse)

# Assuming you have downloaded the AL files
# (e.g., "al2018_rv.csv", "al2018.xlsx") from
#  2014-2019 into your documents folder.
set_ipeds_data_folder(path.expand('~'))

ALData <-
  academic_libraries_load_file(year=2014) \%>\%
    union_all(academic_libraries_load_file(year=2015)) \%>\%
    union_all(academic_libraries_load_file(year=2016)) \%>\%
    union_all(academic_libraries_load_file(year=2017)) \%>\%
    union_all(academic_libraries_load_file(year=2018)) \%>\%
    union_all(academic_libraries_load_file(year=2019))
}
\seealso{
academic_libraries_pivot_by_collection, academic_libraries_pivot_by_circ, academic_libraries_pivot_by_expenditure
}
\author{
Blake Madden
}
