% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_pivot_by_circ}
\alias{academic_libraries_pivot_by_circ}
\title{Pivots an Academic Libraries dataset into a circulation dataset.}
\usage{
academic_libraries_pivot_by_circ(data)
}
\arguments{
\item{data}{An academic libraries file loaded from `academic_libraries_load_file`.}
}
\description{
Creates a circulation column (by school and year),
             with a categorical column specifying the circulation type.
}
\note{
It is required that the column names in the input are in short form.
      In other words, do not call `academic_libraries_expand_colnames` on the dataset
      before calling this.
}
