% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_pivot_by_circ_and_collection}
\alias{academic_libraries_pivot_by_circ_and_collection}
\title{Pivots an Academic Libraries dataset into a circulation and collection dataset.}
\usage{
academic_libraries_pivot_by_circ_and_collection(
  data,
  includeGroupedZScores = F
)
}
\arguments{
\item{data}{An academic libraries dataset loaded from `academic_libraries_load_file`.}

\item{includeGroupedZScores}{TRUE to include collection and circulation z-scores, grouped by
academic year. This is useful for comparing collection vs. circulation
in a more standardized way.}
}
\description{
Creates circulation and collection columns (by school and year),
             with a categorical column specifying the media.
}
\note{
It is required that the column names in the input are in short form.
      In other words, do not call `academic_libraries_expand_colnames` on the dataset
      before calling this.

      Also, the media and circulation types are simplified down to "Physical" and "Electronic."
}
