% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_pivot_by_collection}
\alias{academic_libraries_pivot_by_collection}
\title{Pivots an Academic Libraries (wide) dataset into a collections dataset.}
\usage{
academic_libraries_pivot_by_collection(data)
}
\arguments{
\item{data}{An academic libraries dataset loaded from `academic_libraries_load_file`.}
}
\description{
Creates a count column for all media (by school and year),
             with a categorical column specifying the media type.
}
\note{
It is required that the column names in the input are in short form.
      In other words, do not call `academic_libraries_expand_colnames` on the dataset
      before calling this.
}
\examples{
\dontrun{
library(tidyverse)

# Assuming you have downloaded the AL files
# (e.g., "al2018_rv.csv", "al2018.xlsx") from
#  2014-2019 into your documents folder.
set_ipeds_data_folder(path.expand('~'))
  
ALMediaData <-
  academic_libraries_load_file(year=2014,
                               institutions=oh_community_colleges) \%>\%
    bind_rows(academic_libraries_load_file(year=2015)) \%>\%
    bind_rows(academic_libraries_load_file(year=2016)) \%>\%
    bind_rows(academic_libraries_load_file(year=2017)) \%>\%
    bind_rows(academic_libraries_load_file(year=2018)) \%>\%
    bind_rows(academic_libraries_load_file(year=2019)) \%>\%
    academic_libraries_pivot_by_collection()
}
}
