% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_pivot_by_interlibrary_loan}
\alias{academic_libraries_pivot_by_interlibrary_loan}
\title{Pivots an Academic Libraries dataset into an inter-library loan dataset.}
\usage{
academic_libraries_pivot_by_interlibrary_loan(data)
}
\arguments{
\item{data}{An academic libraries dataset loaded from `academic_libraries_load_file`.}
}
\description{
Creates an inter-library column (by school and year),
             with a categorical column specifying the loan type (i.e., outgoing or received).
}
\note{
It is required that the column names in the input are in short form.
      In other words, do not call `academic_libraries_expand_colnames` on the dataset
      before calling this.
}
