% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcademicLibraries.R
\name{academic_libraries_pivot_by_staff}
\alias{academic_libraries_pivot_by_staff}
\title{Pivots an Academic Libraries dataset into a staff dataset.}
\usage{
academic_libraries_pivot_by_staff(data)
}
\arguments{
\item{data}{An academic libraries dataset loaded from `academic_libraries_load_file`.}
}
\description{
Creates a staff dataset (by school and year),
             with a categorical column specifying the employee type (e.g., librarian, student worker, etc.).
}
\note{
It is required that the column names in the input are in short form.
      In other words, do not call `academic_libraries_expand_colnames` on the dataset
      before calling this.
}
