% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census.R
\name{census_load_demographic_percentages}
\alias{census_load_demographic_percentages}
\title{Summarizes a demographic CSV file from "https://www.census.gov/quickfacts/fact/table".}
\usage{
census_load_demographic_percentages(filePath)
}
\arguments{
\item{filePath}{The path to the census demographic CSV file.}
}
\value{
A summary of the race's and gender's percentages within the area.
}
\description{
This summary will include the race and gender descriptions (which match those from IPEDS),
             a percentage column (in decimal format), and a location column showing the area that
             was included in the CSV file.

             This dataset can be useful for joining with IPEDS data, such as new hire data.
}
\note{
The race labels "unknown" and "nonresident" are not included in census data and will
      not be in the results. (Nonresident technically is included, but overlaps the other
      values, rather than being a distinct classification.)
}
\seealso{
fall_staff_summarize_new_hires_percentages_by_race, fall_staff_summarize_new_hires_percentages_by_gender
}
\author{
Blake Madden
}
