% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cip_codes_2020}
\alias{cip_codes_2020}
\title{CIP Codes 2020}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2848 rows and 8 columns.
}
\source{
<https://nces.ed.gov/ipeds/cipcode/resources.aspx?y=56>
}
\usage{
cip_codes_2020
}
\description{
Classification of Instructional Programs, defined in 2020
}
\keyword{datasets}
