% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Completions.R
\name{completions_bottom_institutions_by_cip_code}
\alias{completions_bottom_institutions_by_cip_code}
\title{Summarizes the n number of colleges that awarded the least awards for each CIP code.}
\usage{
completions_bottom_institutions_by_cip_code(data, n = 1)
}
\arguments{
\item{data}{Data loaded from `completions_load_cip_code_file`.}

\item{n}{The number of top colleges to include for each CIP code.}
}
\description{
Summarizes the n number of colleges that awarded the least awards for each CIP code.
}
\note{
In the case of ties within a CIP code, multiple colleges will be included.
      Also, if a college had reported zero awards for a CIP, then they are not counted
      because they clearly should have deleted the CIP instead of entering zero.
      Instead, we will count the colleges that actually reported a real number of awards.
}
\author{
Blake Madden
}
