% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Completions.R
\name{completions_load_cip_code_file}
\alias{completions_load_cip_code_file}
\title{Loads and pivots a Completions 'a' file, creating "Race" and "Gender" columns and a single award-count column.}
\usage{
completions_load_cip_code_file(year, ..., institutions = NULL, CipLevel = 1)
}
\arguments{
\item{year}{The year in the file name to load.}

\item{...}{Institutions to collapse. This is an argument to `fct_collapse`, which is an
institution and its character vector of satellite institutions to collapse into it.
An example could be `"206604" = c("206613")`, and you can use a variadic list of assignments such as these.
Refer to the documentation for `fct_collapse` for further details.}

\item{institutions}{A vector of college IDs (i.e., the UNITID) to filter on.}

\item{CipLevel}{1 to only show the CIP codes and descriptions based on the first 2 digits of the CIP codes; 2 to get the second level CIPs, 3 to use the full descriptions for the awards.}
}
\value{
The transformed completions file.
}
\description{
Loads a CIP Code Completions file (i.e., the 'a' Completions file, such as c2019_a_rv.csv)
and its respective data dictionary and collapses the award count columns into one column.
This will create additional "Race" and "Gender" categorical columns in the dataset.

Also, CIP, award level, and race descriptions are also joined with the returned dataset.

Finally, an "Academic Year" column will also be added so that the output can be combined with other
years' data.
}
\note{
The '99' CIP Code subtotals will be removed from the output. Also, if the final
version of the completions file is not found, then the provisional version will be used.
}
\examples{
\dontrun{
library(tidyverse)

# Assuming you have downloaded the completion and dictionary files
# (e.g., "c2014_a_rv.csv", "c2014_a.xlsx") from 2014-2019 into
#  your documents folder).
set_ipeds_data_folder(path.expand('~'))
  
completionsData <-
  completions_load_cip_code_file(year=2014) \%>\%
    bind_rows(completions_load_cip_code_file(year=2015)) \%>\%
    bind_rows(completions_load_cip_code_file(year=2016)) \%>\%
    bind_rows(completions_load_cip_code_file(year=2017)) \%>\%
    bind_rows(completions_load_cip_code_file(year=2018)) \%>\%
    bind_rows(completions_load_cip_code_file(year=2019))
}
}
\author{
Blake Madden
}
