% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Completions.R
\name{completions_top_institutions_by_cip_code}
\alias{completions_top_institutions_by_cip_code}
\title{Summarizes the n number of colleges that awarded the most awards for each CIP code.}
\usage{
completions_top_institutions_by_cip_code(data, n = 1)
}
\arguments{
\item{data}{Data loaded from `completions_load_cip_code_file`.}

\item{n}{The number of top colleges to include for each CIP code.}
}
\description{
Summarizes the n number of colleges that awarded the most awards for each CIP code.
}
\note{
In the case of ties within a CIP code, multiple colleges will be included.
}
\author{
Blake Madden
}
