% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FallStaff.R
\name{fall_staff_summarize_new_hires_percentages_by_race}
\alias{fall_staff_summarize_new_hires_percentages_by_race}
\title{Summarizes new hires data into percentage breakdowns by race and institution.}
\usage{
fall_staff_summarize_new_hires_percentages_by_race(data)
}
\arguments{
\item{data}{The new hires data loaded from `fall_staff_load_new_hires_file`.}
}
\value{
The percentages summary.
}
\description{
Summarizes new hires data into percentage breakdowns by race and institution.
}
\author{
Blake Madden
}
