% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finance.R
\name{finance_load_endowments_file}
\alias{finance_load_endowments_file}
\title{Loads and pivots a finance 'a' file, pivoting it into an endowment dataset.}
\usage{
finance_load_endowments_file(year, ..., institutions = NULL)
}
\arguments{
\item{year}{The year in the file name to load.}

\item{...}{Institutions to collapse. This is an argument to `fct_collapse`, which is an
institution and its character vector of satellite institutions to collapse into it.
An example could be `"206604" = c("206613")`, and you can use a variadic list of assignments such as these.
Refer to the documentation for `fct_collapse` for further details.}

\item{institutions}{A vector of college IDs (i.e., the UNITID) to filter on.}
}
\value{
The transformed endowments file.
}
\description{
Loads a finance file (i.e., the 'a' finance file, such as f2019_a_rv.csv)
and converts it into an endowments dataset.

Finally, an "Academic Year" column will also be added so that the output can be combined with other
years' data.
}
\examples{
\dontrun{
library(tidyverse)

# Assuming you have downloaded the finance and dictionary files
# (e.g., "f202223_a_rv.csv", "f202122_a.xlsx") from 2014-2019 into
#  your documents folder).
set_ipeds_data_folder(path.expand('~'))
  
completionsData <-
  finance_load_endowments_file(year=2014) \%>\%
    bind_rows(finance_load_endowments_file(year=2015)) \%>\%
    bind_rows(finance_load_endowments_file(year=2016)) \%>\%
    bind_rows(finance_load_endowments_file(year=2017)) \%>\%
    bind_rows(finance_load_endowments_file(year=2018)) \%>\%
    bind_rows(finance_load_endowments_file(year=2019))
}
}
\author{
Blake Madden
}
