% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstitutionalCharacteristics.R
\name{institution_characteristics_load_directory_info_file}
\alias{institution_characteristics_load_directory_info_file}
\title{Searches for and loads the most recent institutional characteristics directory information file.}
\usage{
institution_characteristics_load_directory_info_file()
}
\value{
The latest IC dataset (if found).
}
\description{
Searches for and loads the most recent institutional characteristics directory information file.
}
\note{
This can be useful for joining with other files (by "UNITID") to add the full
institutions' names, geo-locations, web addresses, etc.
}
\examples{
\dontrun{
library(tidyverse)
library(magrittr)

# Assuming you have downloaded the completion, dictionary, and IC files
# (e.g., "c2014_a_rv.csv", "c2014_a.xlsx", & hd2019.csv) from
#  2014-2019 into into your documents folder.
set_ipeds_data_folder(path.expand('~'))
institutionFile <- institution_characteristics_load_directory_info_file()
  
if (!is.null(institutionFile))
    {
    institutionFile \%<>\%
        dplyr::select(UNITID, INSTNM, CITY, STABBR,
                  WEBADDR, LONGITUD, LATITUDE)
    completionsData <-
      completions_load_cip_code_file(year=2014) \%>\%
      bind_rows(completions_load_cip_code_file(year=2015)) \%>\%
      bind_rows(completions_load_cip_code_file(year=2016)) \%>\%
      bind_rows(completions_load_cip_code_file(year=2017)) \%>\%
      bind_rows(completions_load_cip_code_file(year=2018)) \%>\%
      bind_rows(completions_load_cip_code_file(year=2019))
      
    if (!is.null(completionsData))
        {
        completionsData \%<>\%
            left_join(institutionFile, by = c("UNITID" = "UNITID"))
        }
    }
}
}
\author{
Blake Madden
}
