% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstitutionalCharacteristics.R
\name{institution_characteristics_split_institution_and_campus}
\alias{institution_characteristics_split_institution_and_campus}
\title{Splits institution column into institution and campus columns.}
\usage{
institution_characteristics_split_institution_and_campus(data)
}
\arguments{
\item{data}{An institution directory dataset loaded from `load_institution_directory_info_file`.}
}
\value{
The updated dataset with the new columns.
}
\description{
Splits the `INSTNM` column in an institutional characteristics directory information dataset
             into `Institution` and `Campus` columns.
}
\note{
The `INSTNM` will be removed after the split.
}
\seealso{
load_institution_directory_info_file
}
\author{
Blake Madden
}
