% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstitutionalCharacteristics.R
\docType{data}
\name{oh_community_colleges}
\alias{oh_community_colleges}
\title{A vector containing all Ohio Community College IDs.}
\format{
An object of class \code{character} of length 23.
}
\usage{
oh_community_colleges
}
\description{
The UNITIDs of Ohio Community Colleges.
}
\keyword{datasets}
