% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InstitutionalCharacteristics.R
\name{set_ipeds_data_folder}
\alias{set_ipeds_data_folder}
\title{Sets the folder to read IPEDS data files from.}
\usage{
set_ipeds_data_folder(dataPath)
}
\arguments{
\item{dataPath}{The folder to read files from.}
}
\description{
This should be called before any other function.
}
